@@ -1,32 +1,32 @@
 {
   "private": true,
   "scripts": {
-    "prod": "gulp --production",
-    "dev": "gulp watch"
+    "dev": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
+    "watch": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --watch --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
+    "watch-poll": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --watch --watch-poll --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
+    "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
+    "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
   },
   "devDependencies": {
-    "autoprefixer-loader": "^3.1.0",
-    "babel-core": "^6.7.7",
-    "babel-eslint": "^6.0.4",
-    "babel-loader": "^6.2.4",
+    "axios": "^0.15.3",
     "bootstrap-sass": "^3.3.7",
-    "css-loader": "^0.23.1",
-    "gulp": "^3.9.1",
-    "html-loader": "^0.4.3",
-    "jquery": "^3.1.0",
+    "cross-env": "^3.2.3",
+    "dropzone": "^5.1.1",
+    "jquery": "^3.1.1",
     "laravel-blade-jade": "^2.0.3",
-    "laravel-elixir": "^6.0.0-14",
-    "laravel-elixir-vue-2": "^0.2.0",
+    "laravel-elixir": "^6.0.0-16",
+    "laravel-elixir-jade": "^0.2.4",
+    "laravel-elixir-vue-2": "^0.3.0",
     "laravel-elixir-webpack-official": "^1.0.10",
-    "lodash": "^4.16.2",
-    "node-sass": "^4.5.0",
-    "sass-loader": "^3.2.3",
-    "style-loader": "^0.13.1",
-    "vue": "^2.0.1",
-    "vue-html-loader": "^1.2.2",
-    "vue-loader": "^8.3.1",
-    "vue-resource": "^1.0.3",
-    "vue-style-loader": "^1.0.0",
-    "vuestrap-theme-loader": "^0.1.2"
+    "laravel-mix": "^0.8.3",
+    "lodash": "^4.17.4",
+    "pug": "^2.0.0-rc.4",
+    "vue": "^2.1.10",
+    "vue-resource": "^1.3.4",
+    "vue-router": "^2.7.0",
+    "vuex": "^2.4.0"
+  },
+  "dependencies": {
+    "vue-youtube-embed": "^2.1.2"
   }
 }
